Basic Elements Commit
Project is currently composed of classes:
- IntersectionManager
- RoadSegment
- Road:Roadsegment
- Intersection:Roadsegment
- Path
-Corner

The intersection manager is executed once at the beginning to read the
connections of the intersections and generates a roads for each
connection.

Intersections contain 4 empty objects, one for each corner. The Path
class stores the 4 corners that make a connection (2 of one side of the
start intersection, two on the end intersection). We previously decided
that paths would store Vector2, but passing objects makes it easier to
register dynamic changes (intersection position and dimensions).

TODO:
+ Adapt code so that Intersection accepts a 3 point intersection.
+ Add the Car script and add the GetPosition function of the car.