Added Car functionality, modified base classes
Modified RoadSegment class:
·It has a nextRoadSegment RoadSegment that states the next connection.
- It has an UpdatePath method called by the Car script that generates
the path member of the RoadSegment according to the car state. In the
case of roads, the car state is irrelevant but with intersections, the
directional light and the entry side are required

Added Car script:
- It has a directional light enum that tells the state of the blinking
light. A method register the input of the user to change the directional
lights.
- It has a currentRoadSegment that indicates the road or intersection
that the car is traversing
- It has a method UpdatePosition that asks the currentRoadSegment which
position of the path is it in. If the car if beyond the path length, the
currentRoadSegment is updated.

- Added a GameTester class. The game tester first allows the
IntersectionManager to generate the roads and then it instantiates the
car. This is a provisional class that emulates the game flow.

TODO:
- Improve blinking lights visualization (probably better with rotation)
- Car throws an error if it turns around in an intersection to a side
without exit. Assign a random exit or crash the car.